座標を倍精度小数に変更

+++ src/main/kotlin/com/kamesuta/paintcraft/canvas/CanvasInteraction.kt
@@ -1,7 +1,7 @@
 package com.kamesuta.paintcraft.canvas
 
 import com.kamesuta.paintcraft.frame.FrameRayTraceResult
-import com.kamesuta.paintcraft.util.vec.Vec2i
+import com.kamesuta.paintcraft.util.vec.Vec2d
 import org.bukkit.entity.Player
 
 /**
@@ -11,7 +11,7 @@ import org.bukkit.entity.Player
  * @param player インタラクションを行ったプレイヤー
  */
 data class CanvasInteraction(
-    val uv: Vec2i,
+    val uv: Vec2d,
     val ray: FrameRayTraceResult,
     val player: Player,
 )
